requirements.txt
fastapi
uvicorn
pydantic
responses
groq
python-dotenv
logging
streamlit
requests
uuid